ci: remove downstream slop and point Packit to the package
That whole thing is used to just plainly run `make check` so it doesn't
warrant all that downstream machinery. If it can ever get to the point
where it would be more useful than that it can be reconsidered. Other
than that various projects (including avahi) don't carry that so it's
unclear why nss-mdns should be special in this regard.